refactor: Add QR property to IUsers interface
The code changes in users.interface.ts add a new property, qr, to the IUsers interface. This change enhances the functionality of the Users module by including the QR code as part of the user data structure. The commit message follows the established convention of using the "refactor" type for code changes that improve code structure or organization.