feat: Add QR code support to InsertUser method
The code changes in users.service.ts modify the InsertUser method to include the QR code parameter. This change enhances the functionality of the Users module by allowing the insertion of users with a QR code. The commit message follows the established convention of using the "feat" type for new feature additions.