feat: Add QR code support to ResetPoints method
The code changes in points.controller.ts modify the ResetPoints method to support resetting points for a given QR code. This change enhances the functionality of the Points module by allowing the reset of points using a QR code. The commit message follows the established convention of using the "feat" type for new feature additions.