feat: Add DeleteUser method to UsersService
The code changes in users.service.ts add a new method, DeleteUser, to the UsersService. This method handles the deletion of users by accepting the email as a parameter and deleting the corresponding user record from the database. This change enhances the functionality of the Users module by providing the necessary method for deleting users.