feat: Add findOne method to AttendanceService
The code changes in attendance.service.ts add a new method called `findOne` to the AttendanceService. This method allows retrieving attendance records for a specific user identified by a QR code. This change enhances the functionality of the AttendanceService by providing a way to retrieve attendance information for individual users. The commit message follows the established convention of using the "feat" type for new feature additions.