feat: Add update-user endpoint to UsersController
The code changes in users.controller.ts add a new endpoint called 'update-user' to the UsersController. This endpoint allows updating the role of a user identified by their ID. This change enhances the functionality of the UsersController by providing a way to update user roles. The commit message follows the established convention of using the "feat" type for new feature additions.