chore: Update npm dependencies
This commit updates the npm dependencies in the client/package.json file. It adds the following packages:
- @radix-ui/react-dialog@1.1.1
- @radix-ui/react-popover@1.1.1
- @radix-ui/react-select@2.1.1
- @radix-ui/react-separator@1.1.0
- @tanstack/react-query@5.52.2
- @tanstack/react-table@8.20.5
- cmdk@1.0.0

This chore improves the project by keeping the dependencies up to date with the latest stable versions. The commit message follows the established convention of using the "chore" type for general maintenance tasks.