feat: Update npm dependencies
Update npm dependencies to the latest stable versions. This includes updating axios to version 1.7.5 and zod to version 3.23.8. This update ensures that the application is using the most up-to-date and secure versions of these dependencies. The commit message follows the established convention of using the "feat" type for new feature additions.