feat: Add CreateFirstTimeUser endpoint to UsersController
This commit adds a new endpoint called CreateFirstTimeUser to the UsersController. The endpoint allows the creation of a first-time user by accepting the user's ID number, name, degree, email, and password. It internally generates a QR code using the QrService and inserts the user and QR data into the database using the UsersService. If any error occurs during the process, it throws an HttpException with an appropriate error message. This change enhances the application by providing a convenient way to create first-time users with associated QR codes.