feat: Add QR dialog to DataTableToolbar
This commit adds a QR dialog to the DataTableToolbar component. The QR dialog allows users to add a QR code by filling out the required fields such as ID number, name, and course. This change enhances the functionality of the DataTableToolbar by providing a convenient way to add QR codes for users.