refactor: Update QR columns in DataTable
Refactor the qrcolumns in qr-columns.tsx to update the accessor keys and cell rendering. The "id" accessor key has been changed to "idNumber" and the "course" accessor key has been changed to "degree". This change ensures that the column data is correctly displayed in the DataTable and improves the consistency of the column names.