feat: Add DialogImage component
Add a new component, DialogImage, to handle displaying images in a dialog. This component includes a trigger element and a content section. The trigger element can be any React node and is used to open the dialog. The content section can contain any React node and is displayed inside the dialog. This feature enhances the modularity and reusability of the codebase by providing a standardized way to display images in a dialog.